+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.1.4 Dec 2 2020
+- Add ability to specify session run hash (gorarakelyan)
+- Initialize repo if it was empty when opening session (gorarakelyan)
+- Add validation of map artifact parameters (gorarakelyan)
+
 ## 2.1.3 Nov 24 2020
 - Support comparison of list type contexts (gorarakelyan)